backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
python-multipart>=0.0.6
httpx>=0.25.0
lunar-python>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
